Split ComponentRequestHandlerTest into its own file
